Pacakager Controller

+++ EPaczucha.core/DTO/PackagePriceDto.cs
@@ -3,8 +3,8 @@
     public class PackagePriceDto
     {
         public int Id { get; set; }
-        public int Net { get; set; }
+        public decimal Net { get; set; }
         public int VAT { get; set; }
-        public int Gross { get; set; }
+        public decimal Gross { get; set; }
     }
 }
