+++ EPaczucha.database/Entities/PackagePrice.cs
@@ -2,8 +2,8 @@
 {
     public class PackagePrice : BaseEntity
     {
-        public int Net { get; set; }
+        public decimal Net { get; set; }
         public int VAT { get; set; }
-        public int Gross { get; set; }
+        public decimal Gross { get; set; }
     }
 }
